move pathfinder to fileops
