+++ src/Storage.py
@@ -2,14 +2,14 @@ import os
 import boto3
 from dotenv import load_dotenv
 from multiprocessing import Pool
-from FileOps import FileReader
+from FileOps import PathFinder
 
 
 class Uploader:
     def __init__(self):
         load_dotenv()
         self.bucket = os.environ.get('S3_BUCKET')
-        self.reader = FileReader()
+        self.finder = PathFinder()
 
     def upload_file(self, path):
         with open(path, 'rb') as data:
@@ -17,7 +17,7 @@ class Uploader:
             s3.Bucket(self.bucket).put_object(Key=path, Body=data)
 
     def upload_dir(self, path, truncate=False):
-        paths = self.reader.get_all_paths(path, truncate=truncate)
+        paths = self.finder.get_all_paths(path, truncate=truncate)
         with Pool() as p:
             p.map(self.upload_file, paths)
